+++ src/index.html
@@ -127,10 +127,6 @@
     </span>
   </a>
 </template>
-<script src="js/resizer.js"></script>
-<script src="js/upload.js"></script>
-<script src="pictures.js"></script>
-<!---<script src="js/check.js"></script>!-->
-<!--<script src="js/upload-stat.min.js"></script>!-->
+<script src="js/main.js"></script>
 </body>
 </html>
