+++ src/js/main.js
@@ -1,1 +1,5 @@
 'use strict';
+
+require("./resizer")
+require("./upload")
+require("./pictures")
