+++ src/js/resizer.js
@@ -1,6 +1,6 @@
 'use strict';
 
-(function() {
+module.exports = (function() {
   /**
    * @constructor
    * @param {string} image
