+++ src/js/upload.js
@@ -7,7 +7,7 @@
 
 'use strict';
 
-(function() {
+module.exports = (function() {
   /** @enum {string} */
   var FileType = {
     'GIF': '',
